% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ald_description.R
\name{ald_description}
\alias{ald_description}
\title{Description of a generic ald dataset}
\description{
Asset Level Data: Physical asset production and capacity data across a
variety of sectors and technologies, including forward looking projected
additions or retirements.
}
\examples{
library(dplyr)

data_dictionary() \%>\%
  filter(dataset == "ald")
}
\seealso{
\link{data_dictionary}, \link{ald_demo}.

Other generic datasets: 
\code{\link{loanbook_description}}
}
\concept{generic datasets}
