% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ald_scenario_demo.R
\docType{data}
\name{ald_scenario_demo}
\alias{ald_scenario_demo}
\title{Asset level data plus scenario with sector/technology market share values}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26613 rows and 18 columns.
}
\usage{
ald_scenario_demo
}
\description{
This dataset is the starting point of any PACTA analysis,
including the bank's methodology. It can be used for matching as well as for
the analysis itself. The dataset includes the asset level data connected to
a scenario dataset, having prepared technology market share and sector market
share values.

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(ald_scenario_demo)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}}
}
\concept{demo datasets}
\keyword{datasets}
