% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Column definitions of all datasets}
\format{
\code{data_dictionary} is a \link{data.frame} with columns:
\itemize{
\item \code{column} (character): The name of a dataset-column.
\item \code{dataset} (character): The name of a dataset.
\item \code{definition} (character): The definition of a dataset-column.
\item \code{typeof} (character): The result of \code{typeof()}, one of double, integer,
logical, or character.
}
}
\usage{
data_dictionary
}
\description{
This dataset provides metadata about all datasets in the package r2dii.data.
}
\examples{
head(data_dictionary)
}
\concept{meta}
\keyword{datasets}
