% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_intensity_scenario_demo.R
\docType{data}
\name{emissions_intensity_scenario_demo}
\alias{emissions_intensity_scenario_demo}
\title{A climate scenario dataset for demonstration}
\format{
\code{emissions_intensity_scenario_demo} is a \link{data.frame} with columns:
\itemize{
\item \code{region} (character): The region to which the pathway is relevant.
\item \code{scenario} (character): The name of the scenario.
\item \code{scenario_source} (character): The source publication from which the scenario was taken.
\item \code{sector} (character): The sector to which the scenario prescribes a pathway.
\item \code{unit} (character): The unit of the value prescribed by the scenario.
\item \code{value} (double): The value prescribed by the scenario.
\item \code{year} (integer): The year at which the pathway value is prescribed.
}
}
\usage{
emissions_intensity_scenario_demo
}
\description{
Fake climate scenario dataset, prepared for the software PACTA (Paris
Agreement Capital Transition Assessment). It imitates climate scenario data
(e.g. from the International Energy Agency (IEA)) including the change
through time in production across industrial sectors (calculated by
\href{https://2degrees-investing.org/}{2DII}). This dataset reflects the less
granular, global emissions intensity targets.

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(emissions_intensity_scenario_demo)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{ald_scenario_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}}
}
\concept{demo datasets}
\keyword{datasets}
