% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_or_brown.R
\docType{data}
\name{green_or_brown}
\alias{green_or_brown}
\title{Determine if a technology is green or brown}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 3 columns.
}
\usage{
green_or_brown
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

See \code{?increasing_or_decreasing}
}
\keyword{internal}
