% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increasing_or_decreasing.R
\docType{data}
\name{increasing_or_decreasing}
\alias{increasing_or_decreasing}
\title{Determine if a technology is increasing or decreasing}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 3 columns.
}
\usage{
increasing_or_decreasing
}
\description{
This dataset provides a simple lookup table to determine if a technology is
meant to increase or decrease to align with a scenario that predicts a less
than 2 degree temperature rise.
}
\section{Definitions}{

\itemize{
\item \code{increasing_or_decreasing} (character): If the technology is increasing or decreasing, as defined by the Paris-aligned IEA scenarios., * \code{sector} (character): The sector to which the technology belongs., * \code{technology} (character): The technology sub-category within the sector.
}
}

\examples{
head(increasing_or_decreasing)
}
\seealso{
\link{data_dictionary}
}
\concept{miscellaneous datasets}
\keyword{datasets}
