% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loanbook_description.R
\name{loanbook_description}
\alias{loanbook_description}
\title{Description of a generic loanbook dataset}
\description{
Corporate credit portfolio of banks that holds information of all loans given
by a bank including information on the borrower, the sector classification of
the borrower and the size of the loan. The information comes from banks and
is proprietary data.
}
\examples{
library(dplyr)

data_dictionary() \%>\%
  filter(dataset == "loanbook")
}
\seealso{
\link{data_dictionary}

Other generic datasets: 
\code{\link{ald_description}}
}
\concept{generic datasets}
