% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_bridge.R
\docType{data}
\name{nace_classification}
\alias{nace_classification}
\title{Dataset to bridge (translate) common sector-classification codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1571 rows and 11 columns.
}
\usage{
nace_classification
}
\description{
This dataset serves as a translation key between common sector-classification
systems and sectors relevant to the 'PACTA' tool
(\url{https://www.transitionmonitor.com/}).
}
\section{Definitions}{

\itemize{
\item \code{borderline} (logical): Flag indicating if PACTA sector and classification code are a borderline match. The value TRUE indicates that the match is uncertain between the PACTA sector and the classification. The value FALSE indicates that the match is certainly perfect or the classification is certainly out of PACTA's scope., * \code{code} (character): NACE version 2.1 code with top-level letter prepended., * \code{Common_Content} (character): Mapping of common content between versions 2.1 and 2., * \code{LEVEL} (character): Original NACE version 2.1 code level., * \code{Mapping_Type} (character): Explanation of mapping between versions 2.1 and 2., * \code{NACE2_CODE} (character): Original NACE version 2 code., * \code{NACE2_HEADING} (character): Original NACE version 2 description., * \code{NACE21_CODE} (character): Original NACE version 2.1 code., * \code{NACE21_HEADING} (character): Original NACE version 2.1 description., * \code{sector} (character): Associated PACTA sector., * \code{version} (character): Column identifying to which NACE version the code belongs.
}
}

\section{Details}{

Classification datasets help to standardize sector classification codes from
the wild to a relevant subset including 'power', 'oil and gas', 'coal',
'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.
}

\examples{
head(nace_classification)
}
\seealso{
\link{data_dictionary}.

Other datasets for bridging sector classification codes: 
\code{\link{gics_classification}},
\code{\link{isic_classification}},
\code{\link{naics_classification}},
\code{\link{psic_classification}},
\code{\link{sector_classifications}},
\code{\link{sic_classification}}
}
\concept{datasets for bridging sector classification codes}
\keyword{datasets}
