% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_demo.R
\docType{data}
\name{overwrite_demo}
\alias{overwrite_demo}
\title{A demonstration dataset used to overwrite specific entity names or sectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\usage{
overwrite_demo
}
\description{
Fake dataset used to manually link loanbook entities to mismatched asset
level entities.

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(overwrite_demo)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{ald_scenario_demo}},
\code{\link{loanbook_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}}
}
\concept{demo datasets}
\keyword{datasets}
