% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.data-package.R
\docType{package}
\name{r2dii.data-package}
\alias{r2dii.data}
\alias{r2dii.data-package}
\title{r2dii.data: Datasets to Align Financial Markets with Climate Goals}
\description{
These datasets support the implementation in R of
    the software 'PACTA' (Paris Agreement Capital Transition Assessment),
    which is a free tool that calculates the alignment between financial
    assets and climate scenarios (<https://2degrees-investing.org/>).
    Financial institutions use 'PACTA' to study how their capital
    allocation impacts the climate. Because both financial institutions
    and market data providers keep their data private, this package
    provides fake, public data to enable the development and use of
    'PACTA' in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.data}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.data/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson@2degrees-investing.org} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [data contributor]
  \item Klaus Hagedorn \email{klaus@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
