% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_isos.R
\docType{data}
\name{region_isos}
\alias{region_isos}
\title{A dataset outlining various region definitions (using iso codes)}
\format{
\code{region_isos} is a \link{data.frame} with columns:
\itemize{
\item \code{isos} (character): Countries in region, defined by iso code.
\item \code{region} (character): Benchmark region name.
}
}
\usage{
region_isos
}
\description{
These datasets support the implementation in R of the software 'PACTA' (Paris
Agreement Capital Transition Assessment), which is a free tool that
calculates the alignment between financial assets and climate scenarios
(\url{https://2degrees-investing.org/}).

TODO: Add description for family iso code.

This is a dataset useful when mapping regional benchmarks to countries.
}
\examples{
head(region_isos)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}}

Other iso codes: 
\code{\link{iso_codes}}
}
\concept{demo datasets}
\concept{iso codes}
\keyword{datasets}
