% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_demo.R
\docType{data}
\name{scenario_demo}
\alias{scenario_demo}
\title{A climate scenario dataset for demonstration}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1323 rows and 7 columns.}
\usage{
scenario_demo
}
\description{
This is a public dataset that may be used for examples and tests.
}
\examples{
scenario_demo
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}}
}
\concept{demo datasets}
\keyword{datasets}
