% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_demo_2020.R
\docType{data}
\name{scenario_demo_2020}
\alias{scenario_demo_2020}
\title{A prepared climate scenario dataset for demonstration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1323 rows and 7 columns.
}
\usage{
scenario_demo_2020
}
\description{
Fake data representing a prepared climate scenario dataset, to be used
directly in a PACTA analysis. It imitates data from climate scenario
providers such as the International Energy Agency (IEA). Climate scenarios
are prepared internally by 2Dii for analysis by determining growth rates in
comparison to a start year.

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(scenario_demo_2020)
}
\seealso{
\link{data_dictionary}

TODO: format

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}}
}
\concept{demo datasets}
\keyword{datasets}
