% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_classifications.R
\docType{data}
\name{sector_classifications}
\alias{sector_classifications}
\title{A view of available sector classification datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2600 rows and 4 columns.
}
\usage{
sector_classifications
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}. The column \code{code_system} names one of the
classification systems that 2dii uses. All other columns are defined at
\code{\link[=data_dictionary]{data_dictionary()}}.
}
\description{
A view of available sector classification datasets
}
\examples{
sector_classifications

dd <- data_dictionary()
dd[dd$dataset == "sector_classifications", ]
}
\seealso{
\code{\link[=data_dictionary]{data_dictionary()}}.

Other datasets for bridging sector classification codes: 
\code{\link{classification_bridge}}
}
\concept{datasets for bridging sector classification codes}
\keyword{datasets}
