% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ald_demo.R
\docType{data}
\name{ald_demo}
\alias{ald_demo}
\title{An asset level dataset for demonstration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17368 rows and 14 columns.
}
\usage{
ald_demo
}
\description{
Fake data about physical assets (e.g. wind turbine power plant
capacities) used to assess the climate alignment of financial portfolios.
It imitates data from market-intelligence databases.

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(ald_demo)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_scenario_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}}
}
\concept{demo datasets}
\keyword{datasets}
