% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ald_demo.R
\docType{data}
\name{ald_demo}
\alias{ald_demo}
\title{An asset level dataset for demonstration}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17368 rows and 13 columns.}
\usage{
ald_demo
}
\description{
This is a public dataset that may be used for examples and tests.
}
\examples{
ald_demo
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{data_dictionary}()},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
\keyword{datasets}
