% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_bridge.R
\docType{data}
\name{classification_bridge}
\alias{classification_bridge}
\alias{isic_classification}
\alias{nace_classification}
\alias{naics_classification}
\title{Datasets to bridge (translate) common sector-classification codes}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 768 rows and 4 columns.}
\usage{
isic_classification

nace_classification

naics_classification
}
\value{
A \link[dplyr:tibble]{dplyr::tibble}.
}
\description{
These datasets help to bridge (translate) common sector classification codes
from the wild to codes we use in 2dii such as 'power', 'oil and gas', 'coal',
'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.
}
\examples{
isic_classification
nace_classification
naics_classification
}
\seealso{
\code{\link[=data_dictionary]{data_dictionary()}}.

Other datasets for bridging sector classification codes: 
\code{\link{sector_classification_df}()}
}
\concept{datasets for bridging sector classification codes}
\keyword{datasets}
