% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{A dataset that defines the columns of r2dii datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63 rows and 4 columns.
}
\usage{
data_dictionary
}
\value{
A \link{data.frame} (subclass 'tbl').
}
\description{
A dataset that defines the columns of r2dii datasets
}
\examples{
head(data_dictionary)
}
\seealso{
Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}}
}
\concept{demo datasets}
\keyword{datasets}
