% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Column definitions of all datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 94 rows and 4 columns.
}
\usage{
data_dictionary
}
\description{
This dataset provides metadata about all datasets in the package r2dii.data.
}
\examples{
head(data_dictionary)
}
\concept{meta}
\keyword{datasets}
