% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{A dataset that defines the columns of r2dii datasets}
\format{
\code{data_dictionary} is a \link{data.frame} with columns:
\itemize{
\item \code{column} (character): The name of a dataset-column.
\item \code{dataset} (character): The name of a dataset.
\item \code{definition} (character): The definition of a dataset-column.
\item \code{typeof} (character): The result of \code{typeof()}, one of double, integer,
logical, or character.
}
}
\usage{
data_dictionary
}
\description{
A dataset that defines the columns of r2dii datasets
}
\examples{
head(data_dictionary)
}
\seealso{
Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}}
}
\concept{demo datasets}
\keyword{datasets}
