% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{A dataset that defiles the columns of r2dii datasets}
\usage{
data_dictionary()
}
\value{
A \link[dplyr:tibble]{dplyr::tibble}.
}
\description{
A dataset that defiles the columns of r2dii datasets
}
\section{Extending the data dictionary}{

To extend the output of \code{data_dictionary()} so that it includes
the \code{definition} and \code{typeof} each \code{column} of a new \code{dataset},
you should fill the Google Sheet at \url{http://bit.ly/data_dictionary_template},
then create a new dataset issue at \url{http://bit.ly/new-dataset-issue}.
}

\examples{
data_dictionary()
}
\seealso{
Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
