% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_or_brown.R
\docType{data}
\name{green_or_brown}
\alias{green_or_brown}
\title{Determine if a technology is green or brown}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 3 columns.
}
\usage{
green_or_brown
}
\description{
This dataset provides a simple lookup table to determine if a technology is
meant to increase (green) or decrease (brown) to align with a future global
warming of less than 2 degrees.
}
\section{Definitions}{

\itemize{
\item \code{green_or_brown} (character): If the technology is green (increasing) or brown (decreasing), as defined by the Paris-aligned IEA scenarios., * \code{sector} (character): The sector to which the technology belongs., * \code{technology} (character): The technology sub-category within the sector.
}
}

\examples{
head(green_or_brown)
}
\seealso{
\link{data_dictionary}
}
\concept{miscellaneous datasets}
\keyword{datasets}
