% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes.R
\docType{data}
\name{iso_codes}
\alias{iso_codes}
\title{A dataset linking country names with their ISO code}
\format{
\code{iso_codes} is a \link{data.frame} with columns:
\itemize{
\item \code{country} (character): Country name.
\item \code{country_iso} (character): Corresponding ISO code.
}
}
\usage{
iso_codes
}
\description{
This is a dataset useful when mapping regional benchmarks to countries.

ISO codes datasets use country codes following the ISO standard
(\url{https://www.iso.org/home.html}).
}
\examples{
head(iso_codes)
}
\seealso{
\link{data_dictionary}

Other iso codes: 
\code{\link{region_isos}}
}
\concept{iso codes}
\keyword{datasets}
