% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes.R
\docType{data}
\name{iso_codes}
\alias{iso_codes}
\title{Countries and codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 286 rows and 2 columns.
}
\usage{
iso_codes
}
\description{
This dataset maps countries to codes.

For information about the ISO standard for country codes see
\url{https://www.iso.org/iso-3166-country-codes.html}.
}
\section{Definitions}{

\itemize{
\item \code{country} (character): Country name., * \code{country_iso} (character): Corresponding ISO code.
}
}

\examples{
head(iso_codes)
}
\seealso{
\link{data_dictionary}

Other iso codes: 
\code{\link{region_isos_demo}},
\code{\link{region_isos}}
}
\concept{iso codes}
\keyword{datasets}
