% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes.R
\docType{data}
\name{iso_codes}
\alias{iso_codes}
\title{A dataset linking country names with their ISO code}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 286 rows and 2 columns.}
\usage{
iso_codes
}
\description{
This is a dataset useful when mapping regional benchmarks to countries.
}
\examples{
iso_codes
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
\keyword{datasets}
