% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loanbook_demo.R
\docType{data}
\name{loanbook_demo}
\alias{loanbook_demo}
\title{A loanbook dataset for demonstration}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 320 rows and 19 columns.}
\usage{
loanbook_demo
}
\description{
This is a public dataset that may be used as a template for banks to create
their own loanbook datasets. It is also useful for examples and tests.
}
\examples{
loanbook_demo
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{iso_codes}},
\code{\link{overwrite_demo}},
\code{\link{region_isos}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
\keyword{datasets}
