% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_demo.R
\docType{data}
\name{overwrite_demo}
\alias{overwrite_demo}
\title{A demonstration dataset used to overwrite specific entity names or sectors}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.}
\usage{
overwrite_demo
}
\description{
Used to manually link loanbook entities to mismatched asset level entities.
}
\examples{
overwrite_demo
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{region_isos}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
\keyword{datasets}
