% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.data-package.R
\docType{package}
\name{r2dii.data-package}
\alias{r2dii.data}
\alias{r2dii.data-package}
\title{r2dii.data: Datasets to Measure the Alignment of Corporate Loan Books with Climate Goals}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

These datasets support the implementation in R of the software 'PACTA' (Paris Agreement Capital Transition Assessment), which is a free tool that calculates the alignment between corporate lending portfolios and climate scenarios (\url{https://www.transitionmonitor.com/}). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals. Because both financial institutions and market data providers keep their data private, this package provides fake, public data to enable the development and use of 'PACTA' in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmi-pacta.github.io/r2dii.data/}
  \item \url{https://github.com/RMI-PACTA/r2dii.data}
  \item Report bugs at \url{https://github.com/RMI-PACTA/r2dii.data/issues}
}

}
\author{
\strong{Maintainer}: Jacob Kastl \email{jacob.kastl@gmail.com} (\href{https://orcid.org/0009-0000-8281-8129}{ORCID}) [contractor]

Authors:
\itemize{
  \item Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor, data contributor]
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor, data contributor]
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4banks@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
