% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_isos.R
\docType{data}
\name{region_isos}
\alias{region_isos}
\title{A dataset outlining various region definitions (using iso codes)}
\format{
\code{region_isos} is a \link{data.frame} with columns:
\itemize{
\item \code{isos} (character): Countries in region, defined by iso code.
\item \code{region} (character): Benchmark region name.
}
}
\usage{
region_isos
}
\description{
This is a dataset useful when mapping regional benchmarks to countries.

TODO: Add description for family iso code.
}
\examples{
head(region_isos)
}
\seealso{
\link{data_dictionary}

Other iso codes: 
\code{\link{iso_codes}}
}
\concept{iso codes}
\keyword{datasets}
