% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_isos.R
\docType{data}
\name{region_isos}
\alias{region_isos}
\title{A dataset outlining various region definitions (using iso codes)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1044 rows and 2 columns.}
\usage{
region_isos
}
\description{
This is a dataset useful when mapping regional benchmarks to countries.
}
\examples{
region_isos
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{iso_codes}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{scenario_demo}}
}
\concept{demo datasets}
\keyword{datasets}
