% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_demo.R
\docType{data}
\name{scenario_demo_2020}
\alias{scenario_demo_2020}
\title{A climate scenario dataset for demonstration FIXME documment as other data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1323 rows and 7 columns.
}
\usage{
scenario_demo_2020
}
\description{
A climate scenario dataset for demonstration FIXME documment as other data
}
\examples{
scenario_demo_2020
}
\seealso{
Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}}
}
\concept{demo datasets}
\keyword{datasets}
