% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_demo_2020.R
\docType{data}
\name{scenario_demo_2020}
\alias{scenario_demo_2020}
\title{A prepared climate scenario dataset for demonstration}
\format{
\code{scenario_demo_2020} is a \link{data.frame} with columns:
\itemize{
\item \code{region} (character): The region to which the pathway is relevant.
\item \code{scenario} (character): The name of the scenario.
\item \code{sector} (character): The sector to which the scenario prescribes a
pathway.
\item \code{smsp} (double): Sector market share percentage of the pathway calculated
in 2020.
\item \code{technology} (character): The technology within the sector to which the
scenario prescribes a pathway.
\item \code{tmsr} (double): Technology market share ratio of the pathway calculated
in 2020.
\item \code{year} (integer): The year at which the pathway value is prescribed.
}
}
\usage{
scenario_demo_2020
}
\description{
Fake climate scenario dataset, prepared for the software PACTA (Paris
Agreement Capital Transition Assessment). It imitates climate scenario data
(e.g. from the International Energy Agency (IEA)) including the change
through time in production across industrial sectors (calculated by
\href{https://2degrees-investing.org/}{2DII}).

Demo datasets are synthetic because most financial data is strictly private;
they help to demonstrate and test the implementation in R of 'PACTA'
(\url{https://2degrees-investing.org/resource/pacta/}).
}
\examples{
head(scenario_demo_2020)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}}
}
\concept{demo datasets}
\keyword{datasets}
