% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_classifications.R
\docType{data}
\name{sector_classifications}
\alias{sector_classifications}
\title{A view of available sector classification datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2600 rows and 4 columns.
}
\usage{
sector_classifications
}
\value{
A \link{data.frame} (subclass 'tbl'). The column \code{code_system} names one
of the classification systems that 2dii uses. All other columns are defined
at \link{data_dictionary}.
}
\description{
A view of available sector classification datasets
}
\examples{
head(sector_classifications)

dd <- data_dictionary
head(dd[dd$dataset == "sector_classifications", ])
}
\seealso{
\link{data_dictionary}.

Other datasets for bridging sector classification codes: 
\code{\link{classification_bridge}}
}
\concept{datasets for bridging sector classification codes}
\keyword{datasets}
